Sleep() takes a DWORD
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@9205 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/common/TimeKeeper.cxx
@@ -220,7 +220,7 @@ void TimeKeeper::sleep(float seconds)
 #endif
 #ifdef HAVE_SLEEP
   // equivalent to _sleep() on win32 (not sleep(3))
-  Sleep(seconds * 1000);
+  Sleep((DWORD)(seconds * 1000));
   return;
 #endif
 #ifdef HAVE_SNOOZE
